Implemented admin password for edit/deletion. Created a table for admin info with password. Each editing or deletion of items/category requires admin password. Adding does not require admin password.

+++ controllers/deleteController.js
@@ -1,7 +1,12 @@
 const db = require("../db/queries");
 
 exports.deleteBrandById = async (req, res) => {
-  const brand_id = req.params.brand_id;
-  await db.deleteBrandById(brand_id);
-  res.redirect("/");
+  const adminQuery = await db.getAdminInfo();
+  const adminPassword = adminQuery[0].adminpassword;
+  if (req.body.adminPassword === adminPassword) {
+    const brand_id = req.params.brand_id;
+    await db.deleteBrandById(brand_id);
+    return res.redirect("/");
+  }
+  res.status(400).send("Incorrect Admin Password");
 };
